requirements.txt
# Core dependencies
streamlit==1.45.1
python-dotenv==1.1.0
pandas==2.2.3
numpy==1.26.2

# LangChain and AI
langchain==0.3.25
langchain-openai==0.3.21
langchain-community==0.3.24
langchain-core==0.3.64
openai==1.82.0

# Vector store and embeddings
faiss-cpu==1.11.0
sentence-transformers==4.1.0

# Data processing and utilities
python-dateutil==2.8.2
pytz==2025.2
PyYAML==6.0.2
faker==24.2.0

# Web and API
requests==2.32.3
httpx==0.28.1

# Database
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10

# Development and testing
pytest==8.0.0
black==24.2.0
ruff==0.11.11
